apps/whatsapp-agent/services/paddleocr/requirements.txt
paddleocr==2.7.0.2
Pillow==9.5.0
